+++ fn_absolute.go
@@ -1,0 +1,18 @@
+package main
+
+import "math"
+
+// absolute(seriesList)
+func absolute(e *expr, from, until int32, values map[metricRequest][]*metricData) []*metricData {
+	return forEachSeriesDo(e, from, until, values, func(a *metricData, r *metricData) *metricData {
+		for i, v := range a.Values {
+			if a.IsAbsent[i] {
+				r.Values[i] = 0
+				r.IsAbsent[i] = true
+				continue
+			}
+			r.Values[i] = math.Abs(v)
+		}
+		return r
+	})
+}
