+++ fn_grep.go
@@ -1,0 +1,32 @@
+package main
+
+import "regexp"
+
+// grep(seriesList, pattern)
+func grep(e *expr, from, until int32, values map[metricRequest][]*metricData) []*metricData {
+
+	arg, err := getSeriesArg(e.args[0], from, until, values)
+	if err != nil {
+		return nil
+	}
+
+	pat, err := getStringArg(e, 1)
+	if err != nil {
+		return nil
+	}
+
+	patre, err := regexp.Compile(pat)
+	if err != nil {
+		return nil
+	}
+
+	var results []*metricData
+
+	for _, a := range arg {
+		if patre.MatchString(a.GetName()) {
+			results = append(results, a)
+		}
+	}
+
+	return results
+}
