+++ fn_sqrt.go
@@ -1,0 +1,35 @@
+package main
+
+import (
+	"fmt"
+	"math"
+
+	"github.com/gogo/protobuf/proto"
+)
+
+// squareRoot(seriesList)
+func sqrt(e *expr, from, until int32, values map[metricRequest][]*metricData) []*metricData {
+	arg, err := getSeriesArg(e.args[0], from, until, values)
+	if err != nil {
+		return nil
+	}
+	var results []*metricData
+
+	for _, a := range arg {
+		r := *a
+		r.Name = proto.String(fmt.Sprintf("squareRoot(%s)", a.GetName()))
+		r.Values = make([]float64, len(a.Values))
+		r.IsAbsent = make([]bool, len(a.Values))
+
+		for i, v := range a.Values {
+			if a.IsAbsent[i] {
+				r.Values[i] = 0
+				r.IsAbsent[i] = true
+				continue
+			}
+			r.Values[i] = math.Sqrt(v)
+		}
+		results = append(results, &r)
+	}
+	return results
+}
